Manny's Quiz Game
This is my first solo project I decided to make. With a small amount of help online I was able to loop one of the questions, but overall a harsh game since if you get an answer wrong it terminates the game. Just a game for some of my close friends to try since it questions based of my personal life.